+++ Source/C++/Core/Ap4EsDescriptor.cpp
@@ -167,7 +167,7 @@ AP4_EsDescriptor::Inspect(AP4_AtomInspector& inspector)
 {
     char info[64];
     AP4_FormatString(info, sizeof(info), "size=%d+%d", 
-                     GetHeaderSize(),m_PayloadSize);
+                     (int)GetHeaderSize(),(int)m_PayloadSize);
     inspector.StartElement("[ESDescriptor]", info);
     inspector.AddField("es_id", m_EsId);
     inspector.AddField("stream_priority", m_StreamPriority);
@@ -251,7 +251,7 @@ AP4_EsIdIncDescriptor::Inspect(AP4_AtomInspector& inspector)
 {
     char info[64];
     AP4_FormatString(info, sizeof(info), "size=%d+%d", 
-                     GetHeaderSize(),m_PayloadSize);
+                     (int)GetHeaderSize(),(int)m_PayloadSize);
     inspector.StartElement("[ES_ID_Inc]", info);
     inspector.AddField("track_id", m_TrackId);
     inspector.EndElement();
@@ -298,7 +298,7 @@ AP4_EsIdRefDescriptor::Inspect(AP4_AtomInspector& inspector)
 {
     char info[64];
     AP4_FormatString(info, sizeof(info), "size=%d+%d", 
-                     GetHeaderSize(),m_PayloadSize);
+                     (int)GetHeaderSize(),(int)m_PayloadSize);
     inspector.StartElement("[ES_ID_Ref]", info);
     inspector.AddField("ref_index", m_RefIndex);
     inspector.EndElement();
